requirements/prod.txt
asyncpg==0.29.0
alembic==1.13.0
fastapi==0.110.1
passlib==1.7.4
pathlib==1.0
pydantic[email]==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyjwt==2.8.0
SQLAlchemy==2.0.29
starlette==0.37.2
python-dotenv==0.21.0
psycopg2==2.9.9
uvicorn==0.29.0